+++ wren-ui/src/apollo/server/repositories/index.ts
@@ -18,4 +18,4 @@ export * from './sqlPairRepository';
 export * from './askingTaskRepository';
 export * from './instructionRepository';
 export * from './apiHistoryRepository';
-export * from './dashboardCacheRefreshRepository';
+export * from './dashboardItemRefreshJobRepository';
